[BE-004] Support for PUT requests with attribute type change + SOLID compliance 
PR #3 [BE-000] Rework

+++ MageNet.Persistence/Migrations/MageNetDbContextModelSnapshot.cs
@@ -33,7 +33,7 @@ namespace MageNet.Persistence.Migrations
                         .HasColumnType("nvarchar(100)");
 
                     b.Property<int>("AttributeType")
-                        .HasColumnType("int");
+                        .HasColumnType("tinyint");
 
                     b.Property<Guid>("EntityId")
                         .HasColumnType("uniqueidentifier");
